feat: port to env
