+++ .env
@@ -1,1 +1,2 @@
-PORT="5000"
+PORT="3000"
+HTTP_PORT = "8181"
